% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt.R
\name{querychat_system_prompt}
\alias{querychat_system_prompt}
\title{Create a system prompt for the chat model}
\usage{
querychat_system_prompt(
  df,
  name,
  data_description = NULL,
  extra_instructions = NULL,
  categorical_threshold = 10
)
}
\arguments{
\item{df}{A data frame to generate schema information from.}

\item{name}{A string containing the name of the table in SQL queries.}

\item{data_description}{Optional description of the data, in plain text or Markdown format.}

\item{extra_instructions}{Optional additional instructions for the chat model, in plain text or Markdown format.}

\item{categorical_threshold}{The maximum number of unique values for a text column to be considered categorical.}
}
\value{
A string containing the system prompt for the chat model.
}
\description{
This function generates a system prompt for the chat model based on a data frame's
schema and optional additional context and instructions.
}
