% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querychat.R
\name{querychat_ui}
\alias{querychat_ui}
\alias{querychat_sidebar}
\title{UI components for querychat}
\usage{
querychat_sidebar(id, width = 400, height = "100\%", ...)

querychat_ui(id)
}
\arguments{
\item{id}{The ID of the module instance.}

\item{width}{The width of the sidebar (when using \code{querychat_sidebar}).}

\item{height}{The height of the sidebar (when using \code{querychat_sidebar}).}

\item{...}{Additional arguments passed to \code{bslib::sidebar} (when using \code{querychat_sidebar}).}
}
\value{
A UI object that can be embedded in a Shiny app.
}
\description{
These functions create UI components for the querychat interface.
\code{querychat_ui} creates a basic chat interface, while \code{querychat_sidebar}
wraps the chat interface in a \code{bslib::sidebar} component designed to be used
as the \code{sidebar} argument to \code{bslib::page_sidebar}.
}
